Code cleanup

+++ src/main/java/com/github/klikli_dev/occultism/client/gui/spirit/SpiritGui.java
@@ -120,8 +120,8 @@ public class SpiritGui<T extends SpiritContainer> extends ContainerScreen<T> {
 //region Static Methods
     public static void drawEntityToGui(int posX, int posY, int scale, float mouseX, float mouseY, LivingEntity entity) {
         //From inventory screen
-        float f = (float)Math.atan((double)(mouseX / 40.0F));
-        float f1 = (float)Math.atan((double)(mouseY / 40.0F));
+        float f = (float)Math.atan(mouseX / 40.0F);
+        float f1 = (float)Math.atan(mouseY / 40.0F);
         RenderSystem.pushMatrix();
         RenderSystem.translatef((float)posX, (float)posY, 1050.0F);
         RenderSystem.scalef(1.0F, 1.0F, -1.0F);
